requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
clearml==1.16.5
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
emoji==2.14.0
eval_type_backport==0.2.0
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.10.0
furl==2.1.3
gitdb==4.0.11
GitPython==3.1.43
gradio==5.6.0
gradio_client==1.4.3
greenlet==3.1.1
gritql==0.1.5
grpcio==1.68.0
grpcio-tools==1.68.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
hyperframe==6.0.1
idna==3.10
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.7
langchain-cli==0.0.32
langchain-community==0.3.7
langchain-core==0.3.19
langchain-mongodb==0.2.0
langchain-openai==0.2.9
langchain-text-splitters==0.3.2
langdetect==1.0.9
langserve==0.3.0
langsmith==0.1.144
libmagic==1.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
olefile==0.47
openai==1.55.0
orderedmultidict==1.0.1
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pathlib2==2.3.7.post1
pillow==11.0.0
portalocker==2.10.1
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.10.1
pyparsing==3.2.0
pypdf==5.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic-bin==0.4.14
python-multipart==0.0.12
python-oxmsg==0.0.1
pytz==2024.2
pywin32==308
PyYAML==6.0.2
qdrant-client==1.12.1
RapidFuzz==3.10.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
ruff==0.7.4
safehttpx==0.1.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
sentence-transformers==3.3.1
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
sse-starlette==1.8.2
starlette==0.41.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
tomlkit==0.12.0
torch==2.5.1
tqdm==4.67.0
transformers==4.46.3
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.16.5
unstructured-client==0.27.0
urllib3==2.2.3
uvicorn==0.32.1
webencodings==0.5.1
websockets==12.0
wrapt==1.16.0
yarl==1.17.2
youtube-transcript-api==0.6.3
